Move usage table skeleton into its own component
